fix(bugs): fix u32 overflow in image size calc, silent i32 truncation, pin ort to rc.10
- image_security.rs: cast width/height to u64 before multiplication to prevent
  overflow on large images (e.g. 16384x16384 RGBA wraps u32::MAX)
- models.rs: replace silent `as i32` cast with `i32::try_from().ok().unwrap_or(0)`
  so JSON numbers outside i32 range don't silently corrupt the rank field
- Cargo.toml/lock: pin ort to =2.0.0-rc.10 (exact) to prevent resolving to
  rc.12 which has a compile error on SessionOptionsAppendExecutionProvider_VitisAI

+++ src/core/image_security.rs
@@ -309,7 +309,7 @@ impl ImageSecurityValidator {
             width: img.width(),
             height: img.height(),
             format: format!("{:?}", img.color()),
-            size_bytes: (img.width() * img.height() * (color_depth as u32 / 8) * if has_alpha { 4 } else { 3 }) as usize,
+            size_bytes: (img.width() as u64 * img.height() as u64 * (color_depth as u64 / 8) * if has_alpha { 4 } else { 3 }) as usize,
             has_alpha,
             color_depth,
         }
